Made the login process even more configurable
Instead of just skipping the login form whenever we can get the suer data from the context,
we now provide an extra setting for that. If we configure the tool to get external user data
but not to skip the login form, the latter is prefilled with the external data.

This also happens when the automatic login with the external data fails for some reason.