Create a central way to add new annotations
This repurposes the `createAnnotation` method of `annotationTool`,
which was previously unused. This can now be used in all places
that need to create annotations as a response to some user action.
It also takes care of remembering the last annotation
created in this way, so that we can use it in operations
triggered by things like keyboard shortcuts.