Moved category specific syncing behavior in the proper place
Previously, the custom `AnnotationsSync` passed a parameter to every one of its calls to `toJSON`.
Apart from not honoring the protocol that Backbone specifies for this function (its parameter is an `Object`,
not just a `Boolean` and it normally comes from higher-up calls to functions like `save`, so it contains stuff like
`{ wait: true }`, etc.) this is also only ever needed by `Category`.

This changes `Category#toJSON` to actually work with an object with a custom key to control
its serialization behavior and also overrides `save` in `Category` to always add that key to its `options`,
which are then passed internally to `toJSON`.