Removed unnecessary template handling logic
`{ isTemplate: true }` is never passed to `addCategory`; in general, the UI does not really support category
and scale templates. The `addCopyFromTemplate` functions are not needed anymore, either, because of this.

Reconstructing the `copyUrl` handling of `AnnotationsSync` is also not necessary,
because we can just pass the appropriate attribute (like `category_id` for a category template) as attribute.
This will then be passed to the server by the synchronization layer as POST parameter instead of in the URL.