Factored video related REST handlers into their own class
I think giving each resource its own class that manages it
is a more maintainable way of writing a REST service in this framework.
But in addition to that, doing this refactoring to the video routes
allows us to do certain operations related to that at the central point
where we create this new resource handler.

One could for example control the access to all video related API calls very easily. ;)