Disabled sources of failure in the Gruntfile for now
The `blanket_qunit` task does not work at all and the JSHint task throws lots of errors,
because the code does not adhere to the configured style.
This makes most of the grunt tasks only usable with the `--force` option,
which is unacceptable, because it mgiht hide other potentially legitimate errors.

I also removed the instructions to use this flag from the documentation.