Made strings localizable
This commit also contains some (code) formatting changes to the templates,
that I was to lazy to separate.
It also adds dependencies to `i18next` and `handlebarsHelpers` (for the `t` helper)
to all the modules that needs these.

Note that developer facing strings are not translated!
This includes for example messages displayed on `console`,
but also validation errors in the model code.
These errors should never just be displayed to the user!
Actually the errors returned should probably not even be strings ...

Furthermore, strings which represent symbols, like "<", ">", "+", etc.
are not localized. This is in accordance with us not currently localizing images.
Taking this even further, layouts are also still fixed;
specifically, localizing to right-to-left languages
would take considerably more work on top of this.

All in all this should suffice to localize the app to German
and many other, mostly European languages, but not much else.