Installed `MediaElementPlayer` dependency
Hiding Chromes download button is easy in modern versions of the browser. (See the last commit.)
However, priior to Chrome 58, this can only be done using private, internal CSS selectors.
Doing this correct and portably (across different versions and platforms Chrome runs on)
would probably involve lots of hacks and is actually not what we want in the end.

What is actually preferrable is to have the video player look (mostly) the same across all browsers.
MediaElementPlayer helps with that. If we want it in the future, it could even provide Flash fallbacks.