Stopped using minified jQuery for now
As with Backbone and Underscore before, this library is so prevalent in the stack
that debugging with the minified version is all but impossible.
Preferrably we would use some kind of switching between development and production mode or sourcemaps,
but that is overkill using the current manual dependency management approach.
Besides, we are not all that consistent about this anyway.